@@ -1,32 +1,31 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit, input } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { faPencil, faPlus, faPowerOff, faRotate, IconDefinition } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'app-button',
-  standalone: true,
-  imports: [
-    CommonModule,FontAwesomeModule
-  ],
-  template: `
-  <button type="{{this.type}}" class="{{this.class}}">
+    selector: 'app-button',
+    imports: [
+        CommonModule, FontAwesomeModule
+    ],
+    template: `
+  <button type="{{this.type()}}" class="{{this.class()}}">
     <fa-icon [icon]="faIcon"></fa-icon>
-    {{this.text}}
+    {{this.text()}}
   </button>`,
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ButtonComponent implements OnInit{ 
 
-  @Input() type: string = "button";
-  @Input() class: string = "";
-  @Input() text: string = "";
-  @Input() icon: string = ""; 
+  readonly type = input<string>("button");
+  readonly class = input<string>("");
+  readonly text = input<string>("");
+  readonly icon = input<string>(""); 
 
   public faIcon! : IconDefinition 
 
   public ngOnInit(): void {
-    switch(this.icon){
+    switch(this.icon()){
       case "add":
         this.faIcon = faPlus;
         break;
