@@ -2,13 +2,9 @@ import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
-  selector: 'app-dashboard',
-  standalone: true,
-  imports: [
-    CommonModule,
-  ],
-  template: './dashboard.component.html',
-  styleUrl: './dashboard.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    selector: 'app-dashboard',
+    imports: [CommonModule],
+    template: './dashboard.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DashboardComponent { }
