@@ -2,12 +2,11 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
 
 @Component({
-  selector: 'app-forbidden',
-  standalone: true,
-  imports: [
-    RouterLink,
-  ],
-  templateUrl: './forbidden.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    selector: 'app-forbidden',
+    imports: [
+        RouterLink,
+    ],
+    templateUrl: './forbidden.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ForbiddenComponent { }
